Move HomeDBManager to gen-server/lib/homedb
